requirements.txt
pygame
requests
gTTS
pystray
Pillow
PyQt6

